requirements.txt
python-dotenv
ipykernel

langchain
langchain-community
langchain_core
langchain-text-splitters
langchain-openai
langchain-groq
langchain_huggingface
huggingface_hub
sentence_transformers
langchain_chroma

chromadb
faiss-cpu
duckdb

pypdf
bs4
beautifulsoup4
arxiv
pymupdf
wikipedia
pymupdf

pandas
openai

duckduckgo_search==5.3.1b1
google-api-python-client

psycopg2
mysql-connector-python
SQLAlchemy

validators==0.28.1
youtube_transcript_api
unstructured
pytube
yt_dlp
numexpr

#interface
streamlit
gradio

langserve
uvicorn
fastapi
sse_starlette

#AWS
boto3
awscli

#NVIDIA
openai
langchain_nvidia_ai_endpoints
langchain_community
faiss-cpu
python-dotenv
streamlit
pypdf

#crewai
crewai
crewai_tools
